+++ common/src/main/java/net/bettercombat/mixin/LivingEntityMixin.java
@@ -2,16 +2,18 @@ package net.bettercombat.mixin;
 
 import net.bettercombat.logic.PlayerAttackHelper;
 import net.bettercombat.logic.PlayerAttackProperties;
+import net.bettercombat.logic.knockback.ConfigurableKnockback;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.player.PlayerEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LivingEntity.class)
-public class LivingEntityInject {
+public class LivingEntityMixin implements ConfigurableKnockback {
 
     // FEATURE: Dual wielded attacking - Client side weapon cooldown for offhand
 
@@ -31,5 +33,18 @@ public class LivingEntityInject {
                 cir.cancel();
             }
         }
+    }
+
+    // MARK: ConfigurableKnockback
+    private float customKnockbackMultiplier_BetterCombat = 1;
+
+    @Override
+    public void setKnockbackMultiplier_BetterCombat(float value) {
+        customKnockbackMultiplier_BetterCombat = value;
+    }
+
+    @ModifyVariable(method = "takeKnockback", at = @At("HEAD"), ordinal = 0, argsOnly = true)
+    public double takeKnockback_HEAD_changeStrength(double knockbackStrength) {
+        return knockbackStrength * customKnockbackMultiplier_BetterCombat;
     }
 }