+++ common/src/main/java/net/bettercombat/utils/PatternMatching.java
@@ -1,4 +1,4 @@
-package net.bettercombat.logic;
+package net.bettercombat.utils;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;